Refactor to put common functions into utils.
